Pre release update.
We have not seen each other for so long. Like it was yesterday ... Wait ... ... Never mind ...
In short, I seriously took up the site in order to complete it as soon as possible, and so that everything looked and worked efficiently. Within a couple of days (or to be more precise, then until 05/22/2020) I will bring the site to full readiness. 05/22/2020 I will have the completion of the project, and there we will already see what all these efforts and several sleepless nights cost.
<!---------------->
So what has been added and changed:
- Removed garbage from VScode
- Now the site has been translated into three languages: Russian, Ukrainian and English (Rejoice)
- SCSS broke into a pair of interconnected subfiles to optimize the work.
- Filled in the page "Contacts"
- The "About Us" page appeared and the info component moved there (about the advantages of the company)
<!---------------->
That's all, see you after the site is completed))

May big PP come with you